Assertions have taken a major overhaul in their architecture and are now an independent class, emulation has also changed to accomadate the newer assertion architecture along with Tests no longer depending on the Engine for assertions.
All assertions are now stored within the Assertions class, each
assertion is namespaced to the specific framework it is designed for
which allows it to support emulation directly out of the box. Emulator
object no longer manages assertion tests, rather it passes them off to
the Assertions object as the framework currently in emulation. All
circular object references have also been cleaned up, minus the few that
cannot be avoided due to PHP limitations with closure binding in 5.3.
Tests now accept an Assertions object on construction which is
referenced for all assertions, this allows decoupling from the engine,
Emulation tests follow the same suite and no longer contain duplicate
code for running assertions and rather contain only the emulation
assertion debugging.