Tagging v0.2.2
This introduces the Result object which contains the results of testing,
this is to avoid having to manually calculate tests results.

Calling an invalid assertion now attempts to provide possible matches to
the incorrect assertion.

Any messages encountered during testing are now displayed at testing end
in the CLI output, messages display type are also displayed allowing for
debugging, system and failure messages all to display.

There are many more changes so please refer to the changelog for more
details regarding this release.