Added verbose level for CLI output, modified CLI color output.
The verbose level can be set to three seperate settings. 1 being the
default outputs a easily readable assertion status, 2 outputs the name
of the assertion along with the status and 3 outputs the test assertion
name and namespace arguments provided and the result.

The color output of testing has been updated to disquinish easily from
system, debug, error and messages. The type of message output has also
been modified for the current messages since most were being sent as a
simple message.